# src/main/resources/api-docs.yaml
openapi: 3.0.1
info:
  title: Help API
  version: "1.0"
servers:
  - url: http://192.168.25.107:8081
    description: Generated server url
paths:
  /categories/{pageId}:
    get:
      tags:
        - category-controller
      operationId: getCategories
      parameters:
        - name: pageId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageCategory'
  /categories:
    get:
      tags:
        - category-controller
      operationId: getListCategories
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
    post:
      tags:
        - category-controller
      operationId: addCategory
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Category'
  /categories/category/{categoryId}:
    get:
      tags:
        - category-controller
      operationId: getCategory
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Category'
    post:
      tags:
        - category-controller
      operationId: modifyUnit
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Category'
  /menu/:
    get:
      tags:
        - menu-controller
      operationId: getProducts
      parameters:
        - name: productName
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /menu/{productCode}:
    get:
      tags:
        - menu-controller
      operationId: getProduct
      parameters:
        - name: productCode
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Product'
...
  /products:
    get:
      tags:
        - product-controller
      operationId: getProducts_2
      parameters:
        - name: productName
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: integer
            format: int32
        - name: active
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      tags:
        - product-controller
      operationId: addProduct
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Product'
  /products/product/{productCode}:
    get:
      tags:
        - product-controller
      operationId: getProduct_1
      parameters:
        - name: productCode
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Product'
    post:
      tags:
        - product-controller
      operationId: modifyProduct
      parameters:
        - name: productCode
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Product'
  /products/{pageId}:
    get:
      tags:
        - product-controller
      operationId: getProducts_1
      parameters:
        - name: pageId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: productName
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: integer
            format: int32
        - name: active
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageProduct'
...
  /units/unit/{unitId}:
    get:
      tags:
        - units-controller
      operationId: getUnit
      parameters:
        - name: unitId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Units'
    post:
      tags:
        - units-controller
      operationId: modifyUnit_1
      parameters:
        - name: unitId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Units'
...
components:
  schemas:
    ErrorDetails:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        message:
          type: string
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int32
        category:
          type: string
    PageCategory:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        pageable:
          $ref: '#/components/schemas/Pageable'
        sort:
          $ref: '#/components/schemas/Sort'
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        empty:
          type: boolean
    Pageable:
      type: object
      properties:
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        paged:
          type: boolean
        unpaged:
          type: boolean
        offset:
          type: integer
          format: int64
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productName:
          type: string
        inStockQty:
          type: number
          format: double
        activeStatus:
          type: boolean
        createdAt:
          type: string
          format: date-time
        restaurant:
          type: boolean
        unit:
          $ref: '#/components/schemas/Units'
        category:
          $ref: '#/components/schemas/Category'
    Units:
      type: object
      properties:
        id:
          type: integer
          format: int32
        unitName:
          type: string
...
    PageProduct:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        pageable:
          $ref: '#/components/schemas/Pageable'
        sort:
          $ref: '#/components/schemas/Sort'
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        empty:
          type: boolean
...